# Api/SwaggerJsonSchema.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Api",
    "version": "1.0"
  },
  "paths": {
    "/api/RoadsideAssistant/{Longitude}/{Latitude}/{limit}": {
      "get": {
        "tags": [
          "RoadsideAssistant"
        ],
        "parameters": [
          {
            "name": "longitude",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "latitude",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/RoadsideAssistant/Update": {
      "post": {
        "tags": [
          "RoadsideAssistant"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAssistantRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAssistantRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAssistantRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/RoadsideAssistant/Reserve": {
      "post": {
        "tags": [
          "RoadsideAssistant"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReserveCustomerRoadSideServiceAssistance"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReserveCustomerRoadSideServiceAssistance"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReserveCustomerRoadSideServiceAssistance"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/RoadsideAssistant/Release": {
      "post": {
        "tags": [
          "RoadsideAssistant"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReserveCustomerRoadSideServiceAssistance"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReserveCustomerRoadSideServiceAssistance"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReserveCustomerRoadSideServiceAssistance"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Customer": {
        "type": "object",
        "properties": {
          "membershipId": {
            "type": "string",
            "nullable": true
          },
          "accountNumber": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GeoLocation": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "ReserveCustomerRoadSideServiceAssistance": {
        "type": "object",
        "properties": {
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "location": {
            "$ref": "#/components/schemas/GeoLocation"
          },
          "roadsideServiceAssistantName": {
            "type": "string",
            "nullable": true
          },
          "zip": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateAssistantRequest": {
        "required": [
          "businessName",
          "zip"
        ],
        "type": "object",
        "properties": {
          "businessName": {
            "minLength": 1,
            "type": "string"
          },
          "zip": {
            "maxLength": 5,
            "minLength": 0,
            "type": "string"
          },
          "location": {
            "$ref": "#/components/schemas/GeoLocation"
          }
        },
        "additionalProperties": false
      }
    }
  }
}