# RoadsideAssistant.Data/ApiModel/GeoLocation.cs
namespace RoadsideAssistant.Data.Entities.ApiModel
{
    public class GeoLocation
    {
        public double Latitude { get; set; }

        public double Longitude { get; set; }
    }
}